requirements.txt
# SHA1:7b4caf57e39cd73687d59806ecdea4aece06a726
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
aiohttp==3.8.3
    # via
    #   -r requirements.in
    #   aiohttp-jinja2
aiohttp-jinja2==1.5
    # via -r requirements.in
aiosignal==1.2.0
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==22.1.0
    # via aiohttp
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via aiohttp
cryptography==38.0.1
    # via
    #   -r requirements.in
    #   pyjwt
frozenlist==1.3.1
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.9
    # via gitpython
gitpython==3.1.27
    # via -r requirements.in
idna==3.4
    # via yarl
jinja2==3.1.2
    # via aiohttp-jinja2
markupsafe==2.1.1
    # via jinja2
marshmallow==3.18.0
    # via -r requirements.in
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
packaging==21.3
    # via marshmallow
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.5.0
    # via -r requirements.in
pyparsing==3.0.9
    # via packaging
smmap==5.0.0
    # via gitdb
types-cryptography==3.3.23
    # via pyjwt
yarl==1.8.1
    # via aiohttp
